chore: consolidate schema validation implementations
