+++ packages/stream-handler-common/src/index.ts
@@ -1,1 +1,2 @@
 export * from './anchor-commit-utils.js'
+export * from './schema-utils.js'
